Refactor admin dashboard review components: - Moved IndexDialog to a new directory structure and updated imports. - Created new ApproveDialog, RejectDialog, and RejectReasonDialog components for handling review actions. - Enhanced ReviewDialog to display admin details and manage approval/rejection flow. - Introduced EditStep enum for managing edit states in the store. - Removed obsolete ApproveDialog and RejectDialog components from the previous structure
