+++ src/app/admin/(panel)/dashboard/store.ts
@@ -7,12 +7,27 @@ export enum ReviewStep{
     RejectReason
 }
 
+export enum EditStep{
+    Edit,
+    Success
+}
+
 interface ReviewState{
     step: ReviewStep
     setStep: (step: ReviewStep) => void
 }
 
+interface EditState{
+    step: EditStep
+    setStep: (step: EditStep) => void
+}
+
 export const useStore = create<ReviewState>((set) => ({
     step: ReviewStep.Default,
+    setStep: (step) => set({ step }),
+}))
+
+export const useEditStore = create<EditState>((set) => ({
+    step: EditStep.Edit,
     setStep: (step) => set({ step }),
 }))
